test : 로또 한장 당첨 테스트와 당첨 통계 테스트 분리
